Add of fonts: Ubuntu and Overpass.
Changing the url of the fonts for the local ones.

Creating Dropdown and Footer container and updated of Header, Menu, Paragraph an Section contrainers.

And of course, but not less important, the changes at App.tsx where I changed the texts for the pattern of the previews and the insertion of the Footer.